+++ samples/edd-sample-plugin/edd-sample-plugin.php
@@ -11,19 +11,25 @@ License: MIT
 require_once __DIR__ . '/vendor/autoload.php';
 \WP_Dependency_Installer::instance()->run( __DIR__ );
 
+/**
+ * Load updater.
+ * Must be in main plugin file.
+ *
+ * @return void
+ */
 function edd_test_plugin_updater() {
+	$config = [
+		'type'      => 'plugin', // Declare the type.
+		'file'      => __FILE__,
+		'api_url'   => 'http://eddstore.test', // Site where EDD SL store is located.
+		'item_name' => 'EDD Test Plugin', // Name of plugin.
+		'item_id'   => 11, // ID of the product.
+		'version'   => '1.0', // Current version number.
+		'author'    => 'Andy Fragen', // Author of this plugin.
+		'beta'      => false,
+	];
 	if ( class_exists( 'EDD\\Software_Licensing\\Updater\\Bootstrap' ) ) {
-		( new EDD\Software_Licensing\Updater\Plugin_Updater_Admin(
-			[
-				'file'      => __FILE__,
-				'api_url'   => 'http://eddstore.test',
-				'item_name' => 'EDD Test Plugin',
-				'item_id'   => 11, // ID of the product.
-				'version'   => '1.0', // current version number.
-				'author'    => 'Andy Fragen', // author of this plugin.
-				'beta'      => false,
-			]
-		) )->load_hooks();
+		( new EDD\Software_Licensing\Updater\Init() )->run( $config );
 	}
 }
 add_action( 'plugins_loaded', 'edd_test_plugin_updater' );