+++ CHANGES.md
@@ -6,3 +6,4 @@
 * added initialization and integration for `afragen/translations-updater`
 * update samples to include test for class to prevent fatal when updating Updater
 * add `wp-dependency.json`, `composer require afragen/wp-dependency-installer`, and install code for samples to automatically install EDD SL Updater
+* create `class Init` with universal instantiator function, just add configuration array